requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
websockets==12.0

langchain==0.1.0
langchain-mistralai==0.0.1
langchain-core==0.1.10
langchain-community==0.0.13
langgraph==0.0.20

mistralai==0.0.12

requests==2.31.0
aiohttp==3.9.1
httpx==0.26.0

google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1

# Hugging Face TTS
TTS==0.22.0
piper-tts==1.2.0
transformers==4.36.2
torch==2.1.2
torchaudio==2.1.2

# Audio processing
pydub==0.25.1
soundfile==0.12.1
scipy==1.11.4

# Free TTS alternatives
gTTS==2.5.0
edge-tts==6.1.10

# Video/Animation
manim==0.18.0
moviepy==2.2.1
lottie
Pillow==10.2.0
numpy==1.26.3

python-multipart==0.0.6

pytest==7.4.4
pytest-asyncio==0.23.3